requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
mysqlclient==2.0.3
Pillow==8.2.0
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
